[fix] #23 unitPrice 자료형 변경

+++ module-domain/domain-transports/src/main/java/com/chaing/domain/transports/entity/Transports.java
@@ -45,7 +45,7 @@ public class Transports extends BaseEntity {
     private Integer ownedVehicles;         // 보유 차량 대수
 
     @Column(nullable = false)
-    private Integer unitPrice;             // 운송 단가(박스 + km 당)
+    private Long unitPrice;             // 운송 단가(박스 + km 당)
 
     @Column(nullable = false)
     private LocalDate contractStartDate;  // 계약 시작일
